Move sent-newsletter.tsx into a components folder
