+++ plugins/newsletter-from-post/components/sent-newsletter.tsx
